chore: prepare V.1.3.4
