+++ app/src/main/java/com/kgs/calendar/ui/KgsCalendarApp.kt
@@ -3792,471 +3792,7 @@ private fun LoginScreen(
 }
 
 @Composable
-private fun CalendarShell(
-    state: CalendarUiState,
-    onMenu: () -> Unit,
-    onDateSelected: (LocalDate) -> Unit,
-    onViewSelected: (CalendarViewMode) -> Unit,
-    onMultiDayCountChanged: (Int) -> Unit,
-    onToday: () -> Unit,
-    onSearch: () -> Unit,
-    onTasks: () -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onEventMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onTaskMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onEventMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onTaskMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onSlotSelected: (LocalDate, LocalTime) -> Unit,
-    onAllDaySlotSelected: (LocalDate) -> Unit,
-    draftEvent: DraftEventSelection?,
-    onDraftEventChanged: (DraftEventSelection) -> Unit,
-    onDraftInteraction: () -> Unit,
-    onDraftTap: () -> Unit,
-    timelineBottomInset: Dp,
-    onDetail: (DetailSheet) -> Unit,
-) {
-    val scope = rememberCoroutineScope()
-    val density = LocalDensity.current
-    var monthOverviewOpen by rememberSaveable { mutableStateOf(false) }
-    var yearStripOpen by rememberSaveable { mutableStateOf(false) }
-    var monthOverviewDismissDragPx by remember { mutableFloatStateOf(0f) }
-    var monthOverviewGestureClosing by remember { mutableStateOf(false) }
-    var overviewMonthText by rememberSaveable { mutableStateOf(YearMonth.from(state.selectedDate).toString()) }
-    var monthMorphDayText by rememberSaveable { mutableStateOf(state.selectedDate.toString()) }
-    var monthDayTransitionPending by remember { mutableStateOf(false) }
-    val overviewMonth = remember(overviewMonthText) { YearMonth.parse(overviewMonthText) }
-    val monthMorphDay = remember(monthMorphDayText) { LocalDate.parse(monthMorphDayText) }
-    val isMonthView = state.selectedView == CalendarViewMode.Month
-    // Shared vertical zoom + scroll for the timed views, so pinch level and scroll
-    // position carry across the 3-day <-> 1-day morph instead of resetting.
-    val dayTimeScroll = rememberScrollState()
-    var dayHourHeightDp by rememberSaveable { mutableStateOf(DefaultHourRowHeightDp) }
-    var dayInitialScrollApplied by rememberSaveable { mutableStateOf(false) }
-    var agendaTodayScrollRequest by remember { mutableStateOf(0) }
-    var agendaScrollTargetDate by remember { mutableStateOf<LocalDate?>(null) }
-    // One-shot request to scroll the month list to a specific month. Set ONLY by explicit jumps
-    // (tapping a year in the strip, or the Today button), never by manual scrolling — so the
-    // month view can jump straight there without the scroll<->selectedDate sync fighting it.
-    var monthJumpRequest by remember { mutableStateOf<YearMonth?>(null) }
-    var handledForegroundRecenterSerial by rememberSaveable { mutableStateOf(0) }
-    val today = LocalCalendarTimeSnapshot.current.today
-
-    LaunchedEffect(state.foregroundRecenterSerial) {
-        if (state.foregroundRecenterSerial <= handledForegroundRecenterSerial) return@LaunchedEffect
-        handledForegroundRecenterSerial = state.foregroundRecenterSerial
-        when (state.selectedView) {
-            CalendarViewMode.Agenda -> {
-                agendaScrollTargetDate = today
-                agendaTodayScrollRequest++
-            }
-            CalendarViewMode.Month -> monthJumpRequest = YearMonth.from(today)
-            else -> Unit
-        }
-    }
-
-    LaunchedEffect(state.selectedDate) {
-        overviewMonthText = YearMonth.from(state.selectedDate).toString()
-        if (!monthDayTransitionPending) monthMorphDayText = state.selectedDate.toString()
-    }
-    LaunchedEffect(isMonthView) {
-        if (isMonthView) {
-            monthOverviewOpen = false
-            monthOverviewGestureClosing = false
-            monthOverviewDismissDragPx = 0f
-        } else {
-            yearStripOpen = false
-        }
-    }
-    val monthOverviewVisible = (monthOverviewOpen || monthOverviewGestureClosing) && !isMonthView
-    val monthOverviewHeight by animateDpAsState(
-        targetValue = if (monthOverviewVisible) overviewMonth.overviewPanelHeight(state.firstDayOfWeek) else 0.dp,
-        animationSpec = tween(220, easing = MotionEmphasized),
-        label = "monthOverviewHeight",
-    )
-    val monthOverviewAlpha by animateFloatAsState(
-        targetValue = if (monthOverviewVisible) 1f else 0f,
-        animationSpec = tween(160, easing = MotionStandard),
-        label = "monthOverviewAlpha",
-    )
-    val monthOverviewHeightPx = with(density) { monthOverviewHeight.toPx() }.coerceAtLeast(1f)
-    val monthOverviewVisibleHeight = with(density) {
-        (monthOverviewHeight.toPx() + monthOverviewDismissDragPx).coerceAtLeast(0f).toDp()
-    }
-    val monthDismissThresholdPx = with(density) { 58.dp.toPx() }
-    val monthDismissProgress = (-monthOverviewDismissDragPx / (monthOverviewHeightPx * 0.55f)).coerceIn(0f, 1f)
-
-    fun applyMonthOverviewDismissDrag(deltaY: Float) {
-        monthOverviewGestureClosing = false
-        monthOverviewDismissDragPx = (monthOverviewDismissDragPx + deltaY).coerceIn(-monthOverviewHeightPx, 0f)
-    }
-
-    fun settleMonthOverviewDismissDrag() {
-        val shouldClose = monthOverviewDismissDragPx <= -monthDismissThresholdPx
-        scope.launch {
-            if (shouldClose) monthOverviewGestureClosing = true
-            animate(
-                initialValue = monthOverviewDismissDragPx,
-                targetValue = if (shouldClose) -monthOverviewHeightPx else 0f,
-                animationSpec = tween(
-                    durationMillis = if (shouldClose) 170 else MotionMedium,
-                    easing = if (shouldClose) MotionStandardAccelerate else MotionEmphasized,
-                ),
-            ) { value, _ ->
-                monthOverviewDismissDragPx = value
-            }
-            if (shouldClose) {
-                monthOverviewOpen = false
-                monthOverviewGestureClosing = false
-                monthOverviewDismissDragPx = -monthOverviewHeightPx
-                delay(240)
-                if (!monthOverviewOpen && !monthOverviewGestureClosing) {
-                    monthOverviewDismissDragPx = 0f
-                }
-            }
-        }
-    }
-
-    Column(
-        Modifier
-            .fillMaxSize(),
-    ) {
-        CalendarToolbar(
-            state = state,
-            onMenu = onMenu,
-            onToday = {
-                if (state.selectedView == CalendarViewMode.Agenda) {
-                    agendaScrollTargetDate = today
-                    agendaTodayScrollRequest++
-                }
-                onToday()
-                // In the month grid, also jump the list to the current month so "Today" actually
-                // brings the view back (selectedDate alone won't move the scroll position).
-                if (isMonthView) monthJumpRequest = YearMonth.from(today)
-            },
-            onSearch = onSearch,
-            onTasks = onTasks,
-            monthOverviewOpen = if (isMonthView) yearStripOpen else monthOverviewOpen,
-            onMonthClick = {
-                if (isMonthView) {
-                    yearStripOpen = !yearStripOpen
-                } else {
-                    monthOverviewGestureClosing = false
-                    monthOverviewDismissDragPx = 0f
-                    monthOverviewOpen = !monthOverviewOpen
-                }
-            },
-        )
-        AnimatedVisibility(
-            visible = isMonthView && yearStripOpen,
-            enter = expandVertically(animationSpec = tween(MotionMedium, easing = MotionEmphasized)) + fadeIn(tween(MotionShort)),
-            exit = shrinkVertically(animationSpec = tween(MotionMedium, easing = MotionStandardAccelerate)) + fadeOut(tween(MotionShort)),
-        ) {
-            YearStrip(
-                selectedYear = state.selectedDate.year,
-                onYearSelected = { year ->
-                    onDateSelected(LocalDate.of(year, 1, 1))
-                    monthJumpRequest = YearMonth.of(year, 1)
-                },
-            )
-        }
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(monthOverviewVisibleHeight)
-                .clipToBounds()
-                .graphicsLayer {
-                    alpha = monthOverviewAlpha * (1f - monthDismissProgress * 0.35f)
-                },
-        ) {
-            MonthOverview(
-                month = overviewMonth,
-                state = state,
-                firstDayOfWeek = state.firstDayOfWeek,
-                onDaySelected = { day ->
-                    overviewMonthText = YearMonth.from(day).toString()
-                    if (state.selectedView == CalendarViewMode.Agenda) {
-                        agendaScrollTargetDate = day
-                        agendaTodayScrollRequest++
-                    }
-                    onDateSelected(day)
-                },
-                onMonthSelected = { month ->
-                    overviewMonthText = month.toString()
-                    onDateSelected(month.atDay(1))
-                },
-                onMonthOffset = { offset ->
-                    val next = overviewMonth.plusMonths(offset)
-                    overviewMonthText = next.toString()
-                    onDateSelected(next.atDay(1))
-                },
-                onDismissDrag = ::applyMonthOverviewDismissDrag,
-                onDismissDragEnd = ::settleMonthOverviewDismissDrag,
-                onDismiss = {
-                    monthOverviewGestureClosing = false
-                    monthOverviewDismissDragPx = 0f
-                    monthOverviewOpen = false
-                },
-            )
-        }
-      Box(
-          modifier = Modifier
-              .fillMaxWidth()
-              .weight(1f),
-      ) {
-      SharedTransitionLayout(modifier = Modifier.matchParentSize()) {
-        CompositionLocalProvider(LocalSharedTransitionScope provides this) {
-        AnimatedContent(
-            // Drive the swap off the view GROUP, not the raw view, so Day<->ThreeDay (same
-            // group) never swaps screens — it stays on the timeline and animates column width.
-            targetState = state.selectedView.group(),
-            // Paint the scene area opaquely so the cross-fade between the outgoing and
-            // incoming views never reveals the (lighter) window behind them — that gap is
-            // the brief white flash seen when morphing 1-day <-> 3-day.
-            modifier = Modifier.fillMaxSize().background(MaterialTheme.colorScheme.background),
-            transitionSpec = {
-                // Timeline (day / 3-day) <-> month grid is the container-transform morph: the
-                // selected day's whole block scales between its month cell and the full day via
-                // ScaleToBounds while the rest of each scene fades through. Day<->3-day never
-                // reaches here — it's the same group, handled inside the timeline by animating
-                // the column width. Everything else is a gentle scale-fade.
-                val timelineMonth =
-                    (initialState == CalendarViewGroup.Timeline && targetState == CalendarViewGroup.MonthGrid) ||
-                        (initialState == CalendarViewGroup.MonthGrid && targetState == CalendarViewGroup.Timeline)
-                if (timelineMonth) {
-                    // The tapped day's whole block is the shared element (fades + scales via
-                    // MorphDayEnter/Exit). The REST of the month grid simply cross-fades —
-                    // gradually dissolving out as a day zooms up, and dissolving back in as it
-                    // shrinks down — so nothing snaps and everything fades from/into transparency.
-                    fadeIn(animationSpec = tween(MorphDurationMs, easing = MorphEasing)) togetherWith
-                        fadeOut(animationSpec = tween(MorphDurationMs, easing = MorphEasing))
-                } else {
-                    (
-                        fadeIn(animationSpec = tween(MotionMedium, delayMillis = 90, easing = MotionStandard)) +
-                            scaleIn(initialScale = 0.985f, animationSpec = tween(MotionMedium, delayMillis = 90, easing = MotionStandard))
-                        ) togetherWith (
-                        fadeOut(animationSpec = tween(90, easing = MotionStandardAccelerate)) +
-                            scaleOut(targetScale = 0.995f, animationSpec = tween(90, easing = MotionStandardAccelerate))
-                        )
-                }.using(SizeTransform(clip = false))
-            },
-            label = "calendarViewMorph",
-        ) { group ->
-          val morphScope = this
-          CompositionLocalProvider(LocalMorphAnimatedVisibilityScope provides morphScope) {
-            when (group) {
-                CalendarViewGroup.Timeline -> TimelineView(
-                    state = state,
-                    selectedView = state.selectedView,
-                    onDateSelected = onDateSelected,
-                    onViewSelected = onViewSelected,
-                    onMultiDayCountChanged = onMultiDayCountChanged,
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onEventMoved = onEventMoved,
-                    onTaskMoved = onTaskMoved,
-                    onEventMovedAllDay = onEventMovedAllDay,
-                    onTaskMovedAllDay = onTaskMovedAllDay,
-                    onSlotSelected = onSlotSelected,
-                    onAllDaySlotSelected = onAllDaySlotSelected,
-                    draftEvent = draftEvent,
-                    onDraftEventChanged = onDraftEventChanged,
-                    onDraftInteraction = onDraftInteraction,
-                    onDraftTap = onDraftTap,
-                    timelineBottomInset = timelineBottomInset,
-                    onDetail = onDetail,
-                    timeScroll = dayTimeScroll,
-                    hourHeightDp = dayHourHeightDp,
-                    onHourHeightChange = { dayHourHeightDp = it },
-                    initialTimeScrollApplied = dayInitialScrollApplied,
-                    onInitialTimeScrollApplied = { dayInitialScrollApplied = true },
-                    monthMorphDay = monthMorphDay,
-                )
-                CalendarViewGroup.MonthGrid -> MonthView(
-                    state = state,
-                    // Recenter the selected date when the user SCROLLS to a different month, but not
-                    // when the settled month already matches the selected date. Without this guard,
-                    // composing the month view (e.g. while the 1-day view morphs back into it) fires
-                    // once with the initial month and resets selectedDate to its day 1 — which yanks
-                    // the morph target off the day you came from and onto the 1st, breaking the
-                    // reverse animation for every day except the 1st.
-                    onMonthChanged = { month ->
-                        if (YearMonth.from(state.selectedDate) != month) onDateSelected(month.atDay(1))
-                    },
-                    onOpenDay = { day ->
-                        if (!monthDayTransitionPending) {
-                            monthDayTransitionPending = true
-                            monthMorphDayText = day.toString()
-                            scope.launch {
-                                // Register the tapped month cell's shared-element key before the
-                                // destination scene enters. Without this frame boundary, selecting
-                                // a different day and swapping views can race, leaving no matching
-                                // source bounds and causing an occasional instant full-size frame.
-                                withFrameNanos { }
-                                onDateSelected(day)
-                                withFrameNanos { }
-                                onViewSelected(CalendarViewMode.Day)
-                                monthDayTransitionPending = false
-                            }
-                        }
-                    },
-                    morphDay = monthMorphDay,
-                    jumpRequest = monthJumpRequest,
-                    onJumpConsumed = { monthJumpRequest = null },
-                    onDetail = onDetail,
-                )
-                CalendarViewGroup.Agenda -> AgendaList(state, onTaskStatusChanged, onDetail, agendaTodayScrollRequest, agendaScrollTargetDate)
-                CalendarViewGroup.Tasks -> TaskInbox(state, onTaskStatusChanged, onDetail)
-            }
-          }
-        }
-        }
-      }
-      if (monthOverviewOpen && !isMonthView) {
-          Box(
-              modifier = Modifier
-                  .matchParentSize()
-                  .zIndex(200f),
-          )
-      }
-      }
-    }
-}
-
-@Composable
-private fun CalendarToolbar(
-    state: CalendarUiState,
-    onMenu: () -> Unit,
-    onToday: () -> Unit,
-    onSearch: () -> Unit,
-    onTasks: () -> Unit,
-    monthOverviewOpen: Boolean,
-    onMonthClick: () -> Unit,
-) {
-    val quietInteraction = remember { MutableInteractionSource() }
-    val navBottom = WindowInsets.navigationBars.asPaddingValues().calculateBottomPadding()
-    val statusTop = WindowInsets.statusBars.asPaddingValues().calculateTopPadding()
-    val hasOpenTasks = remember(state.inboxTasks, state.scheduledOpenTasks, state.datedTasks) {
-        (state.inboxTasks + state.scheduledOpenTasks + state.datedTasks)
-            .distinctBy { it.resourceHref }
-            .any { !it.isInactive() }
-    }
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .zIndex(30f)
-            .height(58.dp + statusTop)
-            .background(MaterialTheme.colorScheme.background)
-            .clickable(interactionSource = quietInteraction, indication = null, onClick = {})
-            .padding(start = 4.dp, top = statusTop, end = 8.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(6.dp),
-    ) {
-        IconButton(onClick = onMenu, modifier = Modifier.size(42.dp)) {
-            Icon(Icons.Default.Menu, contentDescription = stringResource(R.string.menu), tint = WarmInk, modifier = Modifier.size(25.dp))
-        }
-        Row(
-            modifier = Modifier
-                .clip(RoundedCornerShape(22.dp))
-                .clickable(onClick = onMonthClick)
-                .padding(horizontal = 4.dp, vertical = 3.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(2.dp),
-        ) {
-            AnimatedContent(
-                // In the month view the grid already labels each month, so the toolbar
-                // shows the year (which changes less often as you scroll).
-                targetState = if (state.selectedView == CalendarViewMode.Month) {
-                    state.selectedDate.year.toString()
-                } else {
-                    state.selectedDate.format(DateTimeFormatter.ofPattern("MMMM", LocalAppLocale.current))
-                },
-                transitionSpec = {
-                    ((slideInVertically(animationSpec = tween(MotionMedium, easing = MotionStandard)) { it / 3 } +
-                        fadeIn(animationSpec = tween(MotionMedium, easing = MotionStandard))) togetherWith
-                        (slideOutVertically(animationSpec = tween(MotionShort, easing = MotionStandardAccelerate)) { -it / 3 } +
-                            fadeOut(animationSpec = tween(MotionShort, easing = MotionStandardAccelerate)))).using(SizeTransform(clip = false))
-                },
-                label = "toolbarMonth",
-            ) { month ->
-                Text(
-                    text = month,
-                    style = MaterialTheme.typography.headlineLarge,
-                    fontWeight = FontWeight.SemiBold,
-                    color = WarmInk,
-                    fontSize = 26.sp,
-                    lineHeight = 30.sp,
-                )
-            }
-            Icon(
-                Icons.Default.KeyboardArrowDown,
-                contentDescription = null,
-                tint = WarmInk,
-                modifier = Modifier
-                    .size(20.dp)
-                    .scale(scaleX = 1f, scaleY = if (monthOverviewOpen) -1f else 1f),
-            )
-        }
-        Spacer(Modifier.weight(1f))
-        IconButton(onClick = onSearch, modifier = Modifier.size(40.dp)) {
-            Icon(Icons.Default.Search, contentDescription = stringResource(R.string.search), tint = WarmInk, modifier = Modifier.size(24.dp))
-        }
-        IconButton(onClick = onToday, modifier = Modifier.size(40.dp)) {
-            TodayDateIcon(day = LocalCalendarTimeSnapshot.current.today.dayOfMonth, modifier = Modifier.size(24.dp))
-        }
-        IconButton(onClick = onTasks, modifier = Modifier.size(40.dp)) {
-            Box(contentAlignment = Alignment.Center) {
-                Icon(Icons.Default.CheckCircle, contentDescription = stringResource(R.string.tasks), tint = WarmInk, modifier = Modifier.size(24.dp))
-                if (hasOpenTasks) {
-                    Box(
-                        modifier = Modifier
-                            .align(Alignment.TopEnd)
-                            .offset(x = 2.dp, y = (-1).dp)
-                            .size(8.dp)
-                            .clip(CircleShape)
-                            .background(WarmBrown)
-                            .border(1.5.dp, MaterialTheme.colorScheme.background, CircleShape),
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TodayDateIcon(day: Int, modifier: Modifier = Modifier) {
-    val iconColor = WarmInk
-    Box(
-        modifier = modifier
-            .drawBehind {
-                val stroke = 1.9.dp.toPx()
-                drawRoundRect(
-                    color = iconColor,
-                    topLeft = Offset(stroke / 2, stroke / 2),
-                    size = Size(size.width - stroke, size.height - stroke),
-                    cornerRadius = CornerRadius(4.5.dp.toPx(), 4.5.dp.toPx()),
-                    style = Stroke(width = stroke),
-                )
-                drawLine(
-                    color = iconColor,
-                    start = Offset(stroke * 2.2f, size.height * 0.28f),
-                    end = Offset(size.width - stroke * 2.2f, size.height * 0.28f),
-                    strokeWidth = stroke,
-                )
-            },
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            text = day.toString(),
-            color = WarmInk,
-            fontSize = 11.sp,
-            lineHeight = 11.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.offset(y = 2.dp),
-        )
-    }
-}
-
-@Composable
-private fun MonthView(
+internal fun MonthView(
     state: CalendarUiState,
     onMonthChanged: (YearMonth) -> Unit,
     onOpenDay: (LocalDate) -> Unit,
@@ -4964,8 +4500,8 @@ private val LocalPendingMutations = compositionLocalOf<List<PendingMutationEntit
  * scopes through every signature. Both are null outside the view-switch container, in
  * which case [morphBounds] is a no-op.
  */
-private val LocalSharedTransitionScope = compositionLocalOf<SharedTransitionScope?> { null }
-private val LocalMorphAnimatedVisibilityScope = compositionLocalOf<AnimatedVisibilityScope?> { null }
+internal val LocalSharedTransitionScope = compositionLocalOf<SharedTransitionScope?> { null }
+internal val LocalMorphAnimatedVisibilityScope = compositionLocalOf<AnimatedVisibilityScope?> { null }
 
 // Multiplier on every view-morph / timeline animation duration. 1 = production speed; raise it
 // (e.g. to 3) to slow the motion down for on-device evaluation. (Search for MorphSlowFactor to
@@ -5081,15 +4617,6 @@ private fun DetailSheet.transitionKey(): String = when (this) {
  * tapped day). Driving [AnimatedContent] off the *group* instead of the raw view is what keeps
  * Day?3-day from cross-fading.
  */
-private enum class CalendarViewGroup { Timeline, MonthGrid, Agenda, Tasks }
-
-private fun CalendarViewMode.group(): CalendarViewGroup = when (this) {
-    CalendarViewMode.Day, CalendarViewMode.ThreeDay -> CalendarViewGroup.Timeline
-    CalendarViewMode.Month -> CalendarViewGroup.MonthGrid
-    CalendarViewMode.Agenda -> CalendarViewGroup.Agenda
-    CalendarViewMode.Tasks -> CalendarViewGroup.Tasks
-}
-
 /**
  * Marks this composable as a shared element keyed by [key]. When the active view changes
  * (e.g. 3-day ? 1-day, or month-cell ? 1-day), the element with the same key in the
@@ -5498,7 +5025,7 @@ private fun OneDayList(state: CalendarUiState, onTaskStatusChanged: (String, Str
 }
 
 @Composable
-private fun AgendaList(
+internal fun AgendaList(
     state: CalendarUiState,
     onTaskStatusChanged: (String, String) -> Unit,
     onDetail: (DetailSheet) -> Unit,
@@ -5686,7 +5213,7 @@ private fun AgendaEventCard(event: EventEntity, onClick: () -> Unit) {
 }
 
 @Composable
-private fun TaskInbox(
+internal fun TaskInbox(
     state: CalendarUiState,
     onTaskStatusChanged: (String, String) -> Unit,
     onDetail: (DetailSheet) -> Unit,
