+++ app/src/main/java/com/kgs/calendar/widget/WidgetMonthRemoteViewsRenderer.kt
@@ -2,7 +2,14 @@ package com.kgs.calendar.widget
 
 import android.content.Context
 import android.os.Bundle
+import android.widget.RemoteViews
 import java.time.ZoneId
+
+internal data class MonthWidgetRenderResult(
+    val views: RemoteViews,
+    val hasCompleteData: Boolean,
+    val signature: String?,
+)
 
 internal class WidgetMonthRemoteViewsRenderer(
     context: Context,
